crc-refactoring
moved crc out of the adc module and moved to the utilities module
