+++ src/test_edgepi/unit_tests/test_adc/test_adc_voltage.py
@@ -1,17 +1,9 @@
 """Unit tests for adc_voltage.py module"""
 
-
-from contextlib import nullcontext as does_not_raise
-
 import pytest
+
 from edgepi.utilities.utilities import bitstring_from_list
 from edgepi.adc.adc_constants import ADCNum
-from edgepi.adc.adc_crc_8_atm import (
-    generate_crc_8_table,
-    check_crc,
-    CRCCheckError,
-    CRC_8_ATM_GEN,
-)
 from edgepi.adc.adc_voltage import (
     _code_to_input_voltage,
     _is_negative_voltage,
@@ -20,7 +12,6 @@ from edgepi.adc.adc_voltage import (
     STEP_DOWN_RESISTOR_1,
     STEP_DOWN_RESISTOR_2
 )
-from edgepi.adc.adc_crc_8_atm import CRC_8_ATM_LUT
 
 # pylint: disable=too-many-lines
 
@@ -112,33 +103,6 @@ def test__adc_voltage_to_input_voltage(code, voltage, num_bytes):
     vin = _code_to_input_voltage(code_uint, V_REF, num_bits)
     assert pytest.approx(_adc_voltage_to_input_voltage(vin, GAIN, OFFSET),0.0001) == \
            voltage * (STEP_DOWN_RESISTOR_1 + STEP_DOWN_RESISTOR_2) / STEP_DOWN_RESISTOR_2
-
-
-@pytest.mark.parametrize(
-    "voltage_bytes, crc_code, err",
-    [
-        ([51, 16, 126, 166], 62, does_not_raise()),
-        ([51, 14, 170, 195], 98, does_not_raise()),
-        ([51, 16, 133, 237], 75, does_not_raise()),
-        ([51, 17, 166, 166], 71, does_not_raise()),
-        ([51, 16, 148, 157], 94, does_not_raise()),
-        ([51, 14, 144, 155], 150, does_not_raise()),
-        ([51, 14, 166, 18], 167, does_not_raise()),
-        ([51, 16, 5, 109], 116, does_not_raise()),
-        ([51, 15, 16, 130], 4, does_not_raise()),
-        ([51, 16, 126, 166], 61, pytest.raises(CRCCheckError)),
-        ([51, 14, 170, 195], 99, pytest.raises(CRCCheckError)),
-        ([51, 16, 133, 237], 70, pytest.raises(CRCCheckError)),
-    ],
-)
-def test_crc_8_atm_adc_1(voltage_bytes, crc_code, err):
-    with err:
-        check_crc(voltage_bytes, crc_code)
-
-
-def test_generate_crc_8_table():
-    assert generate_crc_8_table(CRC_8_ATM_GEN) == CRC_8_ATM_LUT
-
 
 @pytest.mark.parametrize(
     "code, ref_resistance, temp_offset, rtd_conv_constant",
